// src/models/api/Stats.ts
export type Stat = {
  id: number;
  speedMinRegular: number;
  speedMaxRegular: number;
  speedMinMaximum: number;
  speedMaxMaximum: number;
  agilityRegular: number;
  agilityMaximum: number;
  firepower: string;
  maneuverRookieRegular: number;
  maneuverRookieMaximum: number;
  maneuverRegularRegular: number;
  maneuverRegularMaximum: number;
  maneuverVeteranRegular: number;
  maneuverVeteranMaximum: number;
}